feat!: dissolve overlays, rename dialog components
Move overlay components to proper categories:
- Modal Dialog: overlays → status-and-feedback
- Sheet: overlays → layout
- Tooltip: overlays → content

Rename dialog components for clarity:
- ndd-dialog → ndd-inline-dialog (inline feedback component)
- ndd-modal-dialog → ndd-dialog (conventional dialog window)

BREAKING CHANGE: overlay components moved, ndd-dialog renamed to
ndd-inline-dialog, ndd-modal-dialog renamed to ndd-dialog
