--- src/components/overlays/modal-dialog/ndd-modal-dialog.stories.js
@@ -1,164 +1,0 @@
-import { html, nothing } from 'lit';
-import './ndd-modal-dialog.ts';
-import '../../actions/button/ndd-button.ts';
-import { ICONS } from '../../content/icon/ndd-icon.ts';
-
-/**
- * De Modal Dialog is een modaal venster met overlay backdrop.
- * Gebruik `ndd-dialog` voor een inline variant zonder overlay.
- *
- * ## Gebruik
- * ```html
- * <ndd-modal-dialog
- *   text="Bevestiging vereist"
- *   supporting-text="Dit kan niet ongedaan worden gemaakt."
- * >
- *   <ndd-button slot="actions" variant="primary" text="Bevestig"></ndd-button>
- *   <ndd-button slot="actions" variant="neutral-tinted" text="Annuleer"></ndd-button>
- * </ndd-modal-dialog>
- * ```
- */
-export default {
-	title: 'Components/Overlays/Modal Dialog',
-	component: 'ndd-modal-dialog',
-	tags: ['autodocs'],
-	parameters: {
-		componentSource: {
-			file: 'src/components/overlays/modal-dialog/ndd-modal-dialog.ts',
-			repository: 'https://github.com/MinBZK/storybook',
-		},
-		status: { type: 'stable' },
-	},
-	argTypes: {
-		variant: {
-			control: 'select',
-			options: ['', 'alert'],
-			description: 'Semantische variant — dwingt een icoon en kleur af',
-			table: { defaultValue: { summary: '' } },
-		},
-		iconName: {
-			control: 'select',
-			options: ['', ...ICONS],
-			name: 'icon-name',
-			description: 'Naam van het ndd-icon icoon; afwezig wanneer niet ingesteld',
-		},
-		text: {
-			control: 'text',
-			description: 'Hoofdtekst',
-		},
-		supportingText: {
-			control: 'text',
-			name: 'supporting-text',
-			description: 'Ondersteunende tekst',
-		},
-	},
-	args: {
-		variant: '',
-		iconName: '',
-		text: 'Dialog titel',
-		supportingText: 'Ondersteunende tekst voor aanvullende context.',
-	},
-};
-
-const openNext = (e) => e.currentTarget.nextElementSibling.show();
-
-export const Standaard = (args) => html`
-	<ndd-button
-		variant="primary"
-		text="Open modal dialog"
-		@click=${openNext}
-	></ndd-button>
-	<ndd-modal-dialog
-		variant=${args.variant || nothing}
-		icon-name=${args.iconName || nothing}
-		text=${args.text}
-		supporting-text=${args.supportingText}
-	>
-		<ndd-button
-			slot="actions"
-			variant="primary"
-			text="Bevestig"
-			@click=${(e) => e.target.closest('ndd-modal-dialog').hide()}
-		></ndd-button>
-		<ndd-button
-			slot="actions"
-			variant="neutral-tinted"
-			text="Annuleer"
-			@click=${(e) => e.target.closest('ndd-modal-dialog').hide()}
-		></ndd-button>
-	</ndd-modal-dialog>
-`;
-
-export const ZonderIcoon = () => html`
-	<ndd-button
-		variant="primary"
-		text="Open modal dialog"
-		@click=${openNext}
-	></ndd-button>
-	<ndd-modal-dialog
-		text="Bevestiging vereist"
-		supporting-text="Weet u zeker dat u door wilt gaan? Dit kan niet ongedaan worden gemaakt."
-	>
-		<ndd-button
-			slot="actions"
-			variant="primary"
-			text="Bevestig"
-			@click=${(e) => e.target.closest('ndd-modal-dialog').hide()}
-		></ndd-button>
-		<ndd-button
-			slot="actions"
-			variant="neutral-tinted"
-			text="Annuleer"
-			@click=${(e) => e.target.closest('ndd-modal-dialog').hide()}
-		></ndd-button>
-	</ndd-modal-dialog>
-`;
-ZonderIcoon.parameters = { controls: { disable: true } };
-
-export const MetIcoon = () => html`
-	<ndd-button
-		variant="primary"
-		text="Open modal dialog"
-		@click=${openNext}
-	></ndd-button>
-	<ndd-modal-dialog
-		icon-name="check-mark-circle"
-		text="Succesvol opgeslagen"
-		supporting-text="Uw wijzigingen zijn opgeslagen."
-	>
-		<ndd-button
-			slot="actions"
-			variant="primary"
-			text="Sluiten"
-			@click=${(e) => e.target.closest('ndd-modal-dialog').hide()}
-		></ndd-button>
-	</ndd-modal-dialog>
-`;
-MetIcoon.parameters = { controls: { disable: true } };
-
-export const Alert = () => html`
-	<ndd-button
-		variant="primary"
-		text="Open modal dialog"
-		@click=${openNext}
-	></ndd-button>
-	<ndd-modal-dialog
-		variant="alert"
-		text="Niet opgeslagen"
-		supporting-text="Als u doorgaat gaan uw wijzigingen verloren."
-	>
-		<ndd-button
-			slot="actions"
-			variant="primary"
-			text="Doorgaan"
-			@click=${(e) => e.target.closest('ndd-modal-dialog').hide()}
-		></ndd-button>
-		<ndd-button
-			slot="actions"
-			variant="neutral-tinted"
-			text="Annuleer"
-			@click=${(e) => e.target.closest('ndd-modal-dialog').hide()}
-		></ndd-button>
-	</ndd-modal-dialog>
-`;
-Alert.parameters = { controls: { disable: true } };
