+++ src/components/status-and-feedback/dialog/ndd-dialog.stories.js
@@ -1,13 +1,11 @@
 import { html, nothing } from 'lit';
 import './ndd-dialog.ts';
 import '../../actions/button/ndd-button.ts';
-import '../../layout/box/ndd-box.ts';
 import { ICONS } from '../../content/icon/ndd-icon.ts';
 
 /**
- * De Dialog is een inline statuscomponent voor lege toestanden, bevestigingen en feedback.
- * Hij vult zijn container en heeft geen overlay of backdrop.
- * Gebruik `ndd-modal-dialog` voor een modaal venster met backdrop.
+ * De Modal Dialog is een modaal venster met overlay backdrop.
+ * Gebruik `ndd-inline-dialog` voor een inline variant zonder overlay.
  *
  * ## Gebruik
  * ```html
@@ -42,18 +40,16 @@ export default {
 			control: 'select',
 			options: ['', ...ICONS],
 			name: 'icon-name',
-			description: 'Naam van het ndd-icon icoon; afwezig wanneer niet ingesteld. Overschrijft het variant-icoon.',
+			description: 'Naam van het ndd-icon icoon; afwezig wanneer niet ingesteld',
 		},
 		text: {
 			control: 'text',
 			description: 'Hoofdtekst',
-			table: { defaultValue: { summary: '' } },
 		},
 		supportingText: {
 			control: 'text',
 			name: 'supporting-text',
-			description: 'Ondersteunende tekst onder de heading',
-			table: { defaultValue: { summary: '' } },
+			description: 'Ondersteunende tekst',
 		},
 	},
 	args: {
@@ -64,61 +60,105 @@ export default {
 	},
 };
 
+const openNext = (e) => e.currentTarget.nextElementSibling.show();
+
 export const Standaard = (args) => html`
+	<ndd-button
+		variant="primary"
+		text="Open modal dialog"
+		@click=${openNext}
+	></ndd-button>
 	<ndd-dialog
 		variant=${args.variant || nothing}
 		icon-name=${args.iconName || nothing}
 		text=${args.text}
 		supporting-text=${args.supportingText}
 	>
-		<ndd-button slot="actions" variant="primary" text="Bevestig"></ndd-button>
-		<ndd-button slot="actions" variant="neutral-tinted" text="Annuleer"></ndd-button>
+		<ndd-button
+			slot="actions"
+			variant="primary"
+			text="Bevestig"
+			@click=${(e) => e.target.closest('ndd-dialog').hide()}
+		></ndd-button>
+		<ndd-button
+			slot="actions"
+			variant="neutral-tinted"
+			text="Annuleer"
+			@click=${(e) => e.target.closest('ndd-dialog').hide()}
+		></ndd-button>
 	</ndd-dialog>
 `;
 
 export const ZonderIcoon = () => html`
+	<ndd-button
+		variant="primary"
+		text="Open modal dialog"
+		@click=${openNext}
+	></ndd-button>
 	<ndd-dialog
 		text="Bevestiging vereist"
 		supporting-text="Weet u zeker dat u door wilt gaan? Dit kan niet ongedaan worden gemaakt."
 	>
-		<ndd-button slot="actions" variant="primary" text="Bevestig"></ndd-button>
-		<ndd-button slot="actions" variant="neutral-tinted" text="Annuleer"></ndd-button>
+		<ndd-button
+			slot="actions"
+			variant="primary"
+			text="Bevestig"
+			@click=${(e) => e.target.closest('ndd-dialog').hide()}
+		></ndd-button>
+		<ndd-button
+			slot="actions"
+			variant="neutral-tinted"
+			text="Annuleer"
+			@click=${(e) => e.target.closest('ndd-dialog').hide()}
+		></ndd-button>
 	</ndd-dialog>
 `;
 ZonderIcoon.parameters = { controls: { disable: true } };
 
 export const MetIcoon = () => html`
+	<ndd-button
+		variant="primary"
+		text="Open modal dialog"
+		@click=${openNext}
+	></ndd-button>
 	<ndd-dialog
 		icon-name="check-mark-circle"
 		text="Succesvol opgeslagen"
 		supporting-text="Uw wijzigingen zijn opgeslagen."
 	>
-		<ndd-button slot="actions" variant="primary" text="Sluiten"></ndd-button>
+		<ndd-button
+			slot="actions"
+			variant="primary"
+			text="Sluiten"
+			@click=${(e) => e.target.closest('ndd-dialog').hide()}
+		></ndd-button>
 	</ndd-dialog>
 `;
 MetIcoon.parameters = { controls: { disable: true } };
 
 export const Alert = () => html`
+	<ndd-button
+		variant="primary"
+		text="Open modal dialog"
+		@click=${openNext}
+	></ndd-button>
 	<ndd-dialog
 		variant="alert"
 		text="Niet opgeslagen"
 		supporting-text="Als u doorgaat gaan uw wijzigingen verloren."
 	>
-		<ndd-button slot="actions" variant="primary" text="Doorgaan"></ndd-button>
-		<ndd-button slot="actions" variant="neutral-tinted" text="Annuleer"></ndd-button>
+		<ndd-button
+			slot="actions"
+			variant="primary"
+			text="Doorgaan"
+			@click=${(e) => e.target.closest('ndd-dialog').hide()}
+		></ndd-button>
+		<ndd-button
+			slot="actions"
+			variant="neutral-tinted"
+			text="Annuleer"
+			@click=${(e) => e.target.closest('ndd-dialog').hide()}
+		></ndd-button>
 	</ndd-dialog>
 `;
 Alert.parameters = { controls: { disable: true } };
-
-export const LegeToestand = () => html`
-	<ndd-box style="height: 400px; display: flex; align-items: center; justify-content: center;">
-		<ndd-dialog
-			icon-name="inbox"
-			text="Geen resultaten"
-			supporting-text="Er zijn geen items gevonden die overeenkomen met uw zoekopdracht."
-		>
-			<ndd-button slot="actions" variant="neutral-tinted" text="Zoekopdracht wissen"></ndd-button>
-		</ndd-dialog>
-	</ndd-box>
-`;
-LegeToestand.parameters = { controls: { disable: true } };
