+++ src/components/status-and-feedback/inline-dialog/ndd-inline-dialog.styles.ts
@@ -1,0 +1,96 @@
+import { css } from 'lit';
+
+/* # ndd-inline-dialog styles */
+
+export const inlineDialogStyles = css`
+
+	/* # Host */
+
+	:host {
+		display: flex;
+		justify-content: center;
+
+		--_icon-color: var(--semantics-content-color);
+	}
+
+	:host([hidden]) {
+		display: none;
+	}
+
+	:host([variant='alert']) {
+		--_icon-color: var(--primitives-color-warning-350);
+	}
+
+
+	/* # Body */
+
+	.inline-dialog__body {
+		display: flex;
+		flex-direction: column;
+		align-items: center;
+		flex-grow: 1;
+		box-sizing: border-box;
+		max-width: var(--primitives-area-480);
+	}
+
+
+	/* # Icon */
+
+	.inline-dialog__icon {
+		display: flex;
+		align-items: center;
+		justify-content: center;
+		width: var(--primitives-space-48);
+		height: var(--primitives-space-48);
+		color: var(--_icon-color);
+		flex-shrink: 0;
+	}
+
+
+	/* # Text */
+
+	.inline-dialog__text {
+		margin: 0;
+		font: var(--primitives-font-body-md-bold-tight);
+		color: var(--semantics-content-color);
+		text-align: center;
+	}
+
+	.inline-dialog__text:focus-visible {
+		box-shadow: none;
+		outline: none;
+	}
+
+
+	/* # Supporting text */
+
+	.inline-dialog__supporting-text {
+		margin: 0;
+		font: var(--primitives-font-body-sm-regular-tight);
+		color: var(--semantics-content-color);
+		text-align: center;
+	}
+
+
+	/* # Content */
+
+	.inline-dialog__content {
+		width: 100%;
+	}
+
+	.inline-dialog__content:not(:has(*)) {
+		display: none;
+	}
+
+
+	/* # Actions */
+
+	.inline-dialog__actions {
+		width: 100%;
+		padding-top: var(--primitives-space-16);
+	}
+
+	.inline-dialog__actions:not(:has(*)) {
+		display: none;
+	}
+`;
