+++ src/components/status-and-feedback/inline-dialog/ndd-inline-dialog.template.ts
@@ -1,0 +1,34 @@
+import { html, nothing } from 'lit';
+import type { NDDInlineDialog } from './ndd-inline-dialog.ts';
+
+export function inlineDialogTemplate(component: NDDInlineDialog) {
+	return html`
+		<div class="inline-dialog__body">
+			${component._resolvedIconName ? html`
+				<div class="inline-dialog__icon">
+					<ndd-icon name=${component._resolvedIconName}></ndd-icon>
+				</div>
+			` : nothing}
+			${component.text ? html`
+				${component.headingLevel === 1 ? html`<h1 class="inline-dialog__text">${component.text}</h1>`
+				: component.headingLevel === 2 ? html`<h2 class="inline-dialog__text">${component.text}</h2>`
+				: component.headingLevel === 3 ? html`<h3 class="inline-dialog__text">${component.text}</h3>`
+				: component.headingLevel === 4 ? html`<h4 class="inline-dialog__text">${component.text}</h4>`
+				: component.headingLevel === 5 ? html`<h5 class="inline-dialog__text">${component.text}</h5>`
+				: component.headingLevel === 6 ? html`<h6 class="inline-dialog__text">${component.text}</h6>`
+				: html`<p class="inline-dialog__text">${component.text}</p>`}
+			` : nothing}
+			${component.supportingText ? html`
+				<p class="inline-dialog__supporting-text">${component.supportingText}</p>
+			` : nothing}
+			<div class="inline-dialog__content">
+				<slot></slot>
+			</div>
+			<div class="inline-dialog__actions">
+				<ndd-button-group flow="vertical">
+					<slot name="actions"></slot>
+				</ndd-button-group>
+			</div>
+		</div>
+	`;
+}
